Move population-sampling related code to its own module
